chore: Move expanded tiles view logic into stackla-widgets
Alll components in ugc-widgets have their own logic encapsulated in the component, except for expanded-tiles.

As a result, it takes a lot of time to add changes in ugc-widgets, as it requires updating of widget-utils package each time we need to update the expanded tile view, which is complex.

This change means we can update the expanded tile view a lot easier.